+++ examples/esm-example/src/main.ts
@@ -1,14 +1,14 @@
 import { Effect } from "effect";
 import * as Http from "effect-http";
-import * as Log from "effect-log";
+import { Log } from "effect-log";
 
 import { ItemRepositoryInMemory } from "./repository.js";
 import { server } from "./server.js";
 
 const program = server.pipe(
   Http.listen({ port: 3000 }),
-  Effect.provideSomeLayer(Log.setPrettyLogger()),
-  Effect.provideLayer(ItemRepositoryInMemory),
+  Effect.provide(Log.setPrettyLogger()),
+  Effect.provide(ItemRepositoryInMemory),
 );
 
 Effect.runPromise(program);